Update struture pages and imports
