+++ src/components/characters/index.ts
