+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./nav-bar";
+export * from "./characters";
