+++ src/pages/home/index.ts
@@ -1,0 +1,1 @@
+export * from "./home";
