+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export * from "./characters";
+export * from "./home";
 export * from "./episodes";
 export * from "./location";
